requirements.txt
Flask == 2.2.2
keras == 2.9.0
tensorflow == 2.9.1
requests == 2.28.1
numpy == 1.23.2
opencv-python == 4.6.0.66
joblib == 1.1.0
pillow == 9.2.0
